Se movio entity.Character a character.Character
