--- src/main/java/org/thegoats/rolgar2/world/WorldBlock.java
@@ -1,10 +1,0 @@
-package org.thegoats.rolgar2.world;
-
-public class WorldBlock {
-    public Block block;
-    public Position position;
-
-    public WorldBlock(Block block, Position position) {
-        this.block = block;
-    }
-}
